update for newer GHCs and ref-fd package
I'm not sure why the type inference changed in 7.8, but GHC was
complaining that it needed impredicative types for the point-free
definition of `runST`. I made it pointier and now it doesn't complain.

ref-fd added a `modifyRef'` in 0.3.0.4, so we now depend on at least
that version and no longer provide a strict version ourselves.

Closes #1 and #2, thanks for the heads up @meditans! :100: